requirements.txt
Django
djangorestframework
psycopg2-binary
boto3